Refactor TopNav and DropdownMenu
