+++ src/components/Landing/components/TopNav/DropdownMenu.tsx
@@ -1,20 +1,20 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faBars } from '@fortawesome/free-solid-svg-icons';
 import React, { useEffect, useRef, useState } from 'react';
-import { Button } from '../types';
+import { Button } from '../../types';
 
 interface Props {
     buttonsList: Button[];
 }
 
 const DropdownMenu: React.FC<Props> = ({ buttonsList }) => {
-    const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-    const nodeRef = useRef<HTMLUListElement>(null);
+    const [isOpen, setOpen] = useState(false);
+    const nodeRef = useRef<HTMLDivElement>(null);
 
     // TODO: Revisit this type
     const handleClickOutside = (event: any) => {
         if (nodeRef.current && !nodeRef.current.contains(event.target)) {
-            return setIsDropdownOpen(false);
+            return setOpen(false);
         }
     };
 
@@ -24,17 +24,14 @@ const DropdownMenu: React.FC<Props> = ({ buttonsList }) => {
     });
 
     const dropdown = (
-        <ul
-            ref={nodeRef}
-            className="absolute -right-3 bg-primary-120 p-3 space-y-2 rounded-b-md shadow-lg"
-        >
+        <ul className="absolute -right-3 bg-primary-120 p-3 space-y-2 rounded-b-md shadow-lg">
             {buttonsList.map((button) => (
                 <li key={button.id}>
                     <button
                         className="text-white"
                         onClick={() => {
                             button.onClick();
-                            setIsDropdownOpen(false);
+                            setOpen(false);
                         }}
                     >
                         {button.name}
@@ -45,16 +42,16 @@ const DropdownMenu: React.FC<Props> = ({ buttonsList }) => {
     );
 
     return (
-        <>
+        <div ref={nodeRef}>
             <button
                 onClick={() => {
-                    setIsDropdownOpen(!isDropdownOpen);
+                    setOpen(!isOpen);
                 }}
             >
                 <FontAwesomeIcon icon={faBars} className="text-white" />
             </button>
-            <div className="relative">{isDropdownOpen && dropdown}</div>
-        </>
+            <div className="relative">{isOpen && dropdown}</div>
+        </div>
     );
 };
